Describe a lot of possible future extensions
I think describing data in evolution steps instead of snapshot
representation is a powerful concept, that could be applied to make
parts of programming.

At the same time, it is better to do one thing well than a dozen things
poorly. Apivolve will focus on the primary use case, network APIs, until
that is working well.